+++ lib/Clarity/compile.ml
@@ -39,6 +39,7 @@ and compile_expression = function
   | Mod (a, b) -> SWIR.Mod (compile_expression a, compile_expression b)
   | Pow (a, b) -> SWIR.Pow (compile_expression a, compile_expression b)
   | Xor (a, b) -> SWIR.Xor (compile_expression a, compile_expression b)
+  | Len expr -> SWIR.Len (compile_expression expr)
 
 and compile_literal = function
   | NoneLiteral -> SWIR.NoneLiteral