+++ lib/Clarity/grammar.ml
@@ -33,6 +33,7 @@ and expression =
   | Mod of expression * expression
   | Pow of expression * expression
   | Xor of expression * expression
+  | Len of expression
 
 and literal =
   | NoneLiteral